fix: Spring Boot 시작 오류 수정 및 뉴스 서비스 아키텍처 리팩토링
- HealthController 중복 빈 충돌 문제 해결
- private 메소드의 @Transactional 문제 수정 (AOP 프록시 이슈)
- 단일 책임 원칙에 따른 NewsCommandService 분리
  * NewsScrapService: 뉴스 스크래핑 담당
  * NewsAiProcessingService: AI 분석 처리 담당
  * NewsPersistenceService: 데이터베이스 저장 담당
  * NewsNotificationService: 중요 뉴스 알림 담당
- NewsCommandService를 오케스트레이터 패턴으로 변경
- 트랜잭션 경계 및 에러 처리 개선
